pyproject.toml
[tool.poetry]
name = "32-django-permissions-prostgres"
version = "0.1.0"
description = ""
authors = ["Chris Reichert <chris.a.reichert@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
